+++ app/assets/v2/js/base.js
@@ -173,6 +173,49 @@ $(document).ready(function() {
     setTimeout(callback, 300);
   });
 
+  // updates expiry timers with countdowns
+  const setDataFormat = function(data) {
+    let str = 'in ';
+
+    if (data.days() > 0)
+      str += data.days() + 'd ';
+    if (data.hours() > 0)
+      str += data.hours() + 'h ';
+    if (data.minutes() > 0)
+      str += data.minutes() + 'm ';
+    if (data.seconds() > 0)
+      str += data.seconds() + 's ';
+
+    return str;
+  };
+
+  const updateTimers = function() {
+    let enterTime = moment();
+
+    $('[data-time]').filter(':visible').each(function() {
+      moment.locale('en');
+      var time = $(this).data('time');
+      var timeFuture = $(this).data('time-future');
+      var timeDiff = moment(time).diff(enterTime, 'sec');
+
+      if (timeFuture && (timeDiff < 0)) {
+        $(this).html('now');
+        $(this).parents('.offer_container').addClass('animate').removeClass('empty');
+        $(this).removeAttr('data-time');
+
+        // let btn = `<a class="btn btn-block btn-gc-blue btn-sm mt-2" href="${timeUrl}">View Action</a>`;
+        // return $(this).parent().next().html(btn);
+        return $(this).parent().append('<div>Refresh to view offer!</div>');
+      }
+
+      const diffDuration = moment.duration(moment(time).diff(moment()));
+
+      $(this).html(setDataFormat(diffDuration));
+    });
+  };
+
+  setInterval(updateTimers, 1000);
+
   $('.faq_item .question').on('click', (event) => {
     $(event.target).parents('.faq_parent').find('.answer').toggleClass('hidden');
     $(event.target).parents('.faq_parent').find('.answer').toggleClass('show');
