+++ app/grants/templates/grants/index.html
@@ -68,22 +68,23 @@ along with this program. If not, see <http://www.gnu.org/licenses/>.
       <div class="container-fluid pt-0 pb-5 grants-matching">
         <div class="container">
           <div class="col-md-12 font-smaller-1 text-center pt-2">
-            <p class="mb-1 font-weight-semibold">
-              $200k Round 4 is has now closed.
+            <h3 class="mb-0 font-weight-semibold">
+              $200k Round 4 has now closed.
+            </h3>
+            <h4 class="mb-3 pt-1 mt-0 font-weight-semibold">
+              Round 5 coming {{next_round_start|date:'Y-m-d'}} 
+                (<span class="m-0" data-time="{{next_round_start|date:'c'}}" data-base_time="{{now|date:'Y-m-d H:i'}}"></span>)
+
+            </h4>
+            <p class="mb-1">
+              The Round 4 match ran from <strong>1/6 through 1/21 at 16:59 PST</strong>
+              <a href="https://gitcoin.co/blog/gitcoin-grants-2020/">More details in the wrap up post here</a>
+            </p>
+            <p class="mb-1 mt-1 pt-1" style="border-top: 1px solid #eee;">
+              <a href="/mission">Gitcoin's mission is to Grow Open Source</a>. Gitcoin Grants is one of our products that fulfills that mission.  It's also the largest pilot for Quadratic Funding, which economist Glen Weyl once called "The mathematically optimal way to fund public goods that communities care about."!
             </p>
             <p class="mb-1">
-              From <strong>1/6 through 1/21 at 16:59 PST</strong> we were matching contributions with quadratic funding.
-              <a href="https://gitcoin.co/blog/gitcoin-grants-2020/">More details here</a>
-            </p>
-            <p class="mb-1">
-              Gitcoin Grants is the largest pilot for Quadratic Funding around!
-              Want to help fund open source in the 'mathematically optimal' way?
-            </p>
-            <p class="hidden">
-              <a href="https://twitter.com/owocki/status/1217993123311177728">Update 2020/1/17 - Statement on Quid Pro Quos on Gitcoin Grants</a>
-            </p>
-            <p class="mb-1">
-              The next CLR Round is  <strong>tentatively scheduled for March</strong>.  Want to be involved?  Click on the 'be a matching partner' button below.
+                Want to be involved?  Want to help fund open source in the 'mathematically optimal' way? Click on the 'be a matching partner' button below.
             </p>
 
             {% if not user.is_authenticated %}
